Next button and results page
Created 'next' button at the end of the answer evaluation function, and once all the questions are up, the user is redirected to the result.html page. chatGPT recommended '.after' to display the next button after the feedback, apparently it's just another DOM manipulation method. Didn't know this was a thing and happy to have acquired a new tool! Started building the results page with the options to take the user back to the first page or reattempt quiz. Super happy with how this is going!