Fixed quiz selection data intake issues
Several issues - mainly with the scripts page fixed to establish proper flow from clicking the start button, to storing the landing page inputs, to redirecting to the quiz page, to fetching the api data using the stored data, then clearing the initial selections. Phew! That was really tough. I think I'm finally done with the functionality portion! now on to styling!